+++ src/main/java/com/asscope/timesheet/web/rest/EmployeeResource.java
@@ -102,49 +102,49 @@ public class EmployeeResource {
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping({"/employees/me/target-work-time/{year}/{month}"})
-    public ResponseEntity<Long> getTargetWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month) {
+    public ResponseEntity<Integer> getTargetWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month) {
     	log.debug("REST request to get worktimeInformation for Employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.getTargetWorkTimeMinutes(principal.getName(), year, month));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping({"/employees/me/work-time/{year}/{month}"})
-    public ResponseEntity<Long> getWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month) {
+    public ResponseEntity<Integer> getWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month) {
     	log.debug("REST request to get worktime in minutes for Employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.getWorkTimeMinutes(principal.getName(), year, month));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping({"/employees/me/target-work-time/{year}/{month}/{day}"})
-    public ResponseEntity<Long> getTargetWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month, @PathVariable("day") Integer day) {
+    public ResponseEntity<Integer> getTargetWorktimeInformation(Principal principal, @PathVariable("year") Integer year, @PathVariable("month") Integer month, @PathVariable("day") Integer day) {
     	log.debug("REST request to get worktimeInformation for Employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.targetWorkTimeMinutes(principal.getName(), year, month, day));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping("/employees/me/target-work-time/{year}/iso-week/{isoWeek}")
-    public ResponseEntity<Long> getTargetWorktimeInformation(Principal principal, @PathVariable("year") int year, @PathVariable("isoWeek") int isoWeek) {
+    public ResponseEntity<Integer> getTargetWorktimeInformation(Principal principal, @PathVariable("year") int year, @PathVariable("isoWeek") int isoWeek) {
     	log.debug("REST request to get weekly worktimeInformation for Employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.weeklyTargetWorktimeMinutes(principal.getName(), year, isoWeek));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping("/employees/me/work-time/{year}/iso-week/{isoWeek}")
-    public ResponseEntity<Long> getWorktimeInformation(Principal principal, @PathVariable("year") int year, @PathVariable("isoWeek") int isoWeek) {
+    public ResponseEntity<Integer> getWorktimeInformation(Principal principal, @PathVariable("year") int year, @PathVariable("isoWeek") int isoWeek) {
     	log.debug("REST request to get weekly worktimeInformation for Employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.weeklyWorkTimeMinutes(principal.getName(), year, isoWeek));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_USER', 'ROLE_ADMIN')")
     @GetMapping("/employees/me/work-time/balance")
-    public ResponseEntity<Long> currentBalance(Principal principal) {
+    public ResponseEntity<Integer> currentBalance(Principal principal) {
     	log.debug("REST request to get current worktime balance for employee : {}", principal.getName());
     	return ResponseEntity.ok().body(employeeService.currentWorktimeBalance(principal.getName()));
     }
     
     @PreAuthorize("hasAnyAuthority('ROLE_ADMIN')")
     @GetMapping("/employees/{id}/work-time/balance")
-    public ResponseEntity<Long> currentBalance(@PathVariable long id) {
+    public ResponseEntity<Integer> currentBalance(@PathVariable long id) {
     	log.debug("REST request to get current worktime balance for employee : {}", id);
     	Optional<Employee> employee = employeeService.findOne(id);
     	if (employee.isPresent()) {
